requirements.txt
flask
torch
torchvision
pillow
scikit-learn
numpy
flask-cors